Фикс игнорирования отсутствия гравитации боргами
